APP: classificação baseada na qtd de doses e distancia pronta

+++ covide_app/lib/models/dados.dart
@@ -5,7 +5,7 @@ class Dados {
   String name;
   String cep;
   List<double> localization;
-  int qtdOfDose;
+  double qtdOfDose;
   double distance;
   String espediente;
   Dados({
